refactor(test_models): Split out CLI tests to cli/test_cli.py
